building/pyproject.toml
[project]
name = "ariel-os-tiny-ml-building"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "tensorflow[and-cuda]==2.21.0",
  "numpy",
  "tensorflow-model-optimization",
  "pyrootutils",
  "iree-compiler",
  "iree-tools-tflite>=20260316.1567",
  "iree-tools-tf>=20260316.1567",
  "iree-runtime>=20241104.1068",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

